Refactor shed's CompressedFile abstraction into galaxy.util.compression_util.
I need to use this from upload code.
